+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/httpResponse/HttpResponseBindingErrorGeneratable.kt
@@ -4,5 +4,5 @@ import software.amazon.smithy.model.shapes.OperationShape
 import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
 
 interface HttpResponseBindingErrorGeneratable {
-    fun renderHttpResponseBinding(ctx: ProtocolGenerator.GenerationContext, op: OperationShape)
+    fun render(ctx: ProtocolGenerator.GenerationContext, op: OperationShape)
 }
