+++ smithy-swift-codegen/src/test/kotlin/mocks/MockHttpResponseBindingErrorGenerator.kt
@@ -1,0 +1,31 @@
+package mocks
+
+import software.amazon.smithy.codegen.core.Symbol
+import software.amazon.smithy.model.shapes.OperationShape
+import software.amazon.smithy.swift.codegen.ServiceGenerator
+import software.amazon.smithy.swift.codegen.SwiftDependency
+import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
+import software.amazon.smithy.swift.codegen.integration.httpResponse.HttpResponseBindingErrorGeneratable
+
+class MockHttpResponseBindingErrorGenerator : HttpResponseBindingErrorGeneratable {
+    override fun render(
+        ctx: ProtocolGenerator.GenerationContext,
+        op: OperationShape
+    ) {
+        val operationErrorName = ServiceGenerator.getOperationErrorShapeName(op)
+        val rootNamespace = ctx.settings.moduleName
+        val httpBindingSymbol = Symbol.builder()
+            .definitionFile("./$rootNamespace/models/$operationErrorName+HttpResponseBinding.swift")
+            .name(operationErrorName)
+            .build()
+
+        ctx.delegator.useShapeWriter(httpBindingSymbol) { writer ->
+            writer.addImport(SwiftDependency.CLIENT_RUNTIME.target)
+            writer.openBlock("extension \$L {", "}", operationErrorName) {
+                writer.openBlock("public init(httpResponse: HttpResponse, decoder: ResponseDecoder? = nil) throws {", "}") {
+                    writer.write("throw ClientError.deserializationFailed(ClientError.dataNotFound(\"Invalid information in current codegen context to resolve the ErrorType\"))")
+                }
+            }
+        }
+    }
+}
